fix: 修复数据库连接池基类子类资源管理混乱(连接信息)导致无法扩展,丢失任务的问题 fix: 修复连接信息类型转换时异常抛出但未捕获的问题.

+++ sylar/basic/config.h
@@ -126,7 +126,7 @@ inline int getJsonValueByPath(const nlohmann::json& json_data, const std::string
         std::string key = path.substr(front_pos, back_pos - front_pos);
         if(std::isdigit(key[0])) {
             int idx = std::stoi(key);
-            if(idx < 0 || idx >= current_node->size()){
+            if(idx < 0 || idx >= static_cast<int>(current_node->size())){
                 return -1;
             }
             current_node = &((*current_node)[idx]);
@@ -145,7 +145,7 @@ inline int getJsonValueByPath(const nlohmann::json& json_data, const std::string
     std::string key = path.substr(front_pos);
     if(std::isdigit(key[0])) {
         int idx = std::stoi(key);
-        if(idx < 0 || idx >= current_node->size()){
+        if(idx < 0 || idx >= static_cast<int>(current_node->size())){
                 return -1;
         }
         current_node = &((*current_node)[idx]);
